+++ desktop/forge.config.ts
@@ -19,8 +19,8 @@ const config: ForgeConfig = {
         config: rendererConfig,
         entryPoints: [
           {
-            html: './src/index.html',
-            js: './src/renderer.ts',
+            html: './src/renderer/index.html',
+            js: './src/renderer/index.ts',
             name: 'main_window',
             preload: {
               js: './src/preload.ts',
