+++ desktop/src/preload.ts
@@ -1,8 +1,17 @@
 // See the Electron documentation for details on how to use preload scripts:
 // https://www.electronjs.org/docs/latest/tutorial/process-model#preload-scripts
-import {Versions} from "./types";
+
+import {Versions} from "./shared";
 
 const {contextBridge, ipcRenderer} = require('electron')
+
+contextBridge.exposeInMainWorld('versions', {
+    node: () => process.versions.node,
+    chrome: () => process.versions.chrome,
+    electron: () => process.versions.electron,
+    ping: () => ipcRenderer.invoke('ping'),
+    // we can also expose variables, not just functions
+} as Versions)
 
 window.addEventListener('DOMContentLoaded', () => {
     const replaceText = (selector: string, text: string) => {
@@ -14,11 +23,3 @@ window.addEventListener('DOMContentLoaded', () => {
         replaceText(`${dependency}-version`, process.versions[dependency])
     }
 })
-
-contextBridge.exposeInMainWorld('versions', {
-    node: () => process.versions.node,
-    chrome: () => process.versions.chrome,
-    electron: () => process.versions.electron,
-    ping: () => ipcRenderer.invoke('ping'),
-    // we can also expose variables, not just functions
-} as Versions)
