+++ desktop/src/renderer/index.css
@@ -5,3 +5,7 @@ body {
   max-width: 38rem;
   padding: 2rem;
 }
+
+.bold {
+  font-weight: bold;
+}
