+++ desktop/src/renderer/index.html
@@ -9,9 +9,9 @@
     <h1>💖 Hello World!</h1>
     <p>Welcome to your Electron application.</p>
     <p>
-      Nous utilisons Node.js <span id="node-version"></span>,
-      Chromium <span id="chrome-version"></span>,
-      et Electron <span id="electron-version"></span>.
+      Nous utilisons Node.js <span id="node-version" class="bold"></span>,
+      Chromium <span id="chrome-version" class="bold"></span>,
+      et Electron <span id="electron-version" class="bold"></span>.
     </p>
     <p id="info"></p>
   </body>
